fix: reduce report token use helper complexity
