+++ .agents/configs/complexity-thresholds-history.md
@@ -162,3 +162,4 @@ and monotonically decreases".
 | 40 | GH#22123/t3379 (PR #22124) | pre-existing drift on main — 38 smells vs threshold 37 (CI reports `Total qlty smells: 38, Threshold: 37`); Qlty Smell Regression (delta-based) skips — PR changes only shell files (setup.sh, deploy-agents-on-merge.sh, test-setup-noninteractive-lock.sh) which qlty does not flag; Qlty Maintainability Smells passes; drift from 35→38 (3 smells) accumulated from intervening merges after t3050 set the threshold; 38 actual + 2 buffer = 40; ratchet-post-merge.yml will reset toward actual after the next smell-reducing merge |
 | 39 | ratchet-post-merge | auto-ratchet after e2a569b ("chore: mark t3390 complete (pr:#22163)"): count 37 + 2 buffer = 39 (previously 40, reduction 1) |
 | 48 | GH#24060 | pre-existing drift on the PR merge ref — 46 smells vs threshold 45. Qlty Smell Regression and `qlty check` passed for the report-render refactor, while the absolute-count threshold job reported no report-render files among the top offenders; 46 actual + 2 buffer = 48. ratchet-post-merge.yml will reset toward actual after the next smell-reducing merge |
+| 53 | GH#24220/GH#24221 | /report-token-use released in v3.20.0 with 3 qlty smells in the initial local reporting helper; post-release main reports 51 smells vs threshold 48. This follow-up splits rendering out for maintainability, but total remains 51. 51 actual + 2 buffer = 53; ratchet back down after helper simplification removes the smells. |
